Porting to LiteLoader, part 2 - everything ported & working, except configs
The configs are currently implemented by copy-pasting the config related
classes from Forge (which are not included in this commit).

The configs will be converted into a JSON system in the next commit(s).

Also, the Reflection and Method Handle stuff is currently broken in a
built mod due to notch obfuscation being used.