Creation of the ShapeSphere
ShapeSphere created. It is a sphere with a customisable radius
Modified ShapeManager to add Normal Sphere support
Modified GuiShapeEditor to create the GUI for the sphere shape.
Modified GuiShapeManager to create a normal sphere on "Add shape". Only a sphere can be selected/created at the moment. Dropdown no longer shows the normal info toggles list.
Modified the en_us language file to add support for the ShapeSphere ("Normal Sphere") and radius