Use the fps value in MinecraftClient instead of updateFps() method
This mod is incompatible with [exordium](https://www.curseforge.com/minecraft/mc-mods/exordium). When installed exordium, the miniHUD huds are not rendered every render tick, and then the fps value of miniHUD is not the real fps of minecraft.

Just use `net.minecraft.client.MinecraftClient#getCurrentFps` to get the correct fps value.